+++ MultiFunPlayer/Common/Keyframe.cs
@@ -3,12 +3,12 @@
 namespace MultiFunPlayer.Common;
 
 [DebuggerDisplay("[{Position}, {Value}]")]
-public record Keyframe(float Position, float Value);
+public record Keyframe(double Position, double Value);
 
 public class KeyframePositionComparer : IComparer<Keyframe>
 {
     public static KeyframePositionComparer Default { get; } = new();
 
     public int Compare(Keyframe x, Keyframe y)
-        => Comparer<float>.Default.Compare(x.Position, y.Position);
+        => Comparer<double>.Default.Compare(x.Position, y.Position);
 }
