+++ MultiFunPlayer/OutputTarget/ViewModels/SerialOutputTargetViewModel.cs
@@ -148,7 +148,7 @@ public class SerialOutputTargetViewModel : ThreadAbstractOutputTarget
             EventAggregator.Publish(new SyncRequestMessage());
 
             var stopwatch = Stopwatch.StartNew();
-            var lastSentValues = DeviceAxis.All.ToDictionary(a => a, _ => float.NaN);
+            var lastSentValues = DeviceAxis.All.ToDictionary(a => a, _ => double.NaN);
             while (!token.IsCancellationRequested && serialPort.IsOpen)
             {
                 stopwatch.Restart();
@@ -160,7 +160,7 @@ public class SerialOutputTargetViewModel : ThreadAbstractOutputTarget
                     Logger.Debug("Received \"{0}\" from \"{1}\"", serialPort.ReadExisting(), SelectedSerialPortDeviceId);
 
                 var dirtyValues = Values.Where(x => DeviceAxis.IsDirty(x.Value, lastSentValues[x.Key]));
-                var commands = DeviceAxis.ToString(dirtyValues, 1000 * stopwatch.ElapsedTicks / (float)Stopwatch.Frequency);
+                var commands = DeviceAxis.ToString(dirtyValues, 1000 * stopwatch.ElapsedTicks / (double)Stopwatch.Frequency);
                 if (serialPort.IsOpen && !string.IsNullOrWhiteSpace(commands))
                 {
                     Logger.Trace("Sending \"{0}\" to \"{1}\"", commands.Trim(), SelectedSerialPortDeviceId);
