+++ MultiFunPlayer/OutputTarget/ViewModels/WebSocketOutputTargetViewModel.cs
@@ -80,7 +80,7 @@ public class WebSocketOutputTargetViewModel : AsyncAbstractOutputTarget
 
                 UpdateValues();
 
-                var commands = DeviceAxis.ToString(Values, 1000 * stopwatch.ElapsedTicks / (float)Stopwatch.Frequency);
+                var commands = DeviceAxis.ToString(Values, 1000 * stopwatch.ElapsedTicks / (double)Stopwatch.Frequency);
                 if (client.State == WebSocketState.Open && !string.IsNullOrWhiteSpace(commands))
                 {
                     Logger.Trace("Sending \"{0}\" to \"{1}\"", commands.Trim(), Uri.ToString());
